Remove duplicate methods that the json gem defines automatically.
I'd forgotten *why* I felt the need to create a `to_json` method.
It wasn't that the default ruby way didn't do it, it was that the
default ruby way was too unpredictable. Instead of fighting the
ecosystem and paper overing where things are rickety, I'm just
going to make my own method where I can take more liberties
and ignore the normal guarantees that JSON generators have to meet.